Refactor: InputView, OutputView 작성
값을 입력받고, 우승자를 출력하는 로직을 RacingCar로부터 분리해서 InputView, OutputView 클래스를 작성함
