+++ src/main/java/car/view/DefaultCarsFormatter.java
@@ -1,4 +1,4 @@
-package car.ui;
+package car.view;
 
 import car.domain.Car;
 import car.domain.CarsFormatter;
@@ -22,6 +22,7 @@ public class DefaultCarsFormatter implements CarsFormatter {
     }
 
     public String formatSingleCar(final Car car) {
-        return car.getName() + NAME_VALUE_JOIN_DELIMITER + DISTANCE_MARKER.repeat(car.getPosition());
+        return car.getName() + NAME_VALUE_JOIN_DELIMITER + DISTANCE_MARKER.repeat(
+            car.getPosition());
     }
 }
